+++ src/onegov.form/onegov/form/parser/__init__.py
@@ -1,3 +1,11 @@
-from onegov.form.parser.core import parse_form
+from onegov.form.parser.core import parse_formcode
+from onegov.form.parser.core import flatten_fieldsets
+from onegov.form.parser.form import parse_form
+from onegov.form.parser.form import WTFormsClassBuilder
 
-__all__ = ['parse_form']
+__all__ = [
+    'flatten_fieldsets',
+    'parse_form',
+    'parse_formcode',
+    'WTFormsClassBuilder'
+]